requirements.txt
alembic==1.12.0
blinker==1.6.2
cachelib==0.10.2
click==8.1.7
dominate==2.8.0
Flask==2.3.3
Flask-Bootstrap==3.3.7.1
Flask-Migrate==4.0.4
Flask-Session==0.5.0
Flask-WTF==1.1.1
greenlet==2.0.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
Pillow==10.0.1
playwright==1.37.0
pyee==9.0.4
Pygments==2.16.1
python-dotenv==1.0.0
six==1.16.0
typing_extensions==4.7.1
visitor==0.1.3
Werkzeug==2.3.7
WTForms==3.0.1
zipp==3.16.2
